Allow anonymous binding in addition to admin binding.
Anonymous binding is a common thing in OpenLDAP, and we should account
for it with this plugin. If binddn and bindpw are missing, assume that
we are attempting to bind anonymously.

If either binddn or bindpw are specified, both parameters are needed -
you cannot do admin binding without both a username and password. The
plugin will croak() if either is missing.